Refactor BlobFileUploader.cs
Fix: App.config, local to write
Clean-up: move using within namespace, remove non-active using
Add: new classes for sub-steps
Add: bool check to new classes to support loop break
