+++ Kiroku/kiroku-logloader/LogUploader/DataModels/InstanceModel.cs
@@ -1,11 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace KLOGLoader
+﻿namespace KLOGLoader
 {
+    using System;
+
     public class InstanceModel
     {
         public string Version { get; set; }
