+++ Kiroku/kiroku-logloader/LogUploader/Program.cs
@@ -1,18 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using Microsoft.Azure;
-using Microsoft.WindowsAzure.Storage;
-using Microsoft.WindowsAzure.Storage.Blob;
-using System.IO;
-using Newtonsoft.Json;
-using System.Collections.Specialized;
-
-// Kiroku
-using Kiroku;
-using System.Configuration;
-
-namespace KLOGLoader
+﻿namespace KLOGLoader
 {
     class Program
     {
