+++ Kiroku/kiroku-logloader/LogUploader/V2DataModel/SQLResponse.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace KLOGLoader
+﻿namespace KLOGLoader
 {
     public class SQLResponse
     {
