+++ Kiroku/kiroku-logloader/LogUploader/V2Global/BlobClient.cs
@@ -1,14 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Microsoft.Azure;
-using Microsoft.WindowsAzure.Storage;
-using Microsoft.WindowsAzure.Storage.Blob;
-
-namespace KLOGLoader
+﻿namespace KLOGLoader
 {
+    using Microsoft.Azure;
+    using Microsoft.WindowsAzure.Storage;
+    using Microsoft.WindowsAzure.Storage.Blob;
+
     public static class BlobClient
     {
         //private static GlobalBlobClient _instance;
