+++ Kiroku/kiroku-logloader/LogUploader/V2Processor/BlobFileCheck.cs
@@ -1,14 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-// Kiroku
-using Kiroku;
-
-namespace KLOGLoader
+﻿namespace KLOGLoader
 {
+    using System;
+    using System.Linq;
+
+    // Kiroku
+    using Kiroku;
+
     public class BlobFileCheck
     {
         public static void Execute()
